pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-radar"
version = "0.1.0"
description = "Daily job digest CLI — scores backend postings against your resume"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.2",
    "jinja2>=3.1.4",
    "typer>=0.12.3",
    "structlog>=24.2.0",
    "selectolax>=0.3.21",
    "feedparser>=6.0.11",
    "anthropic>=0.30.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
job-radar = "radar.__main__:app"

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "respx>=0.21.1",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.12",
]

[tool.hatch.build.targets.wheel]
packages = ["radar"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
